+++ src/Fhir.Anonymizer.Core/Extensions/ElementNodeExtension.cs
@@ -106,64 +106,5 @@ namespace Fhir.Anonymizer.Core.Extensions
         {
             return node.Children("meta").Cast<ElementNode>().FirstOrDefault();
         }
-
-        public static void RemoveNullChildren(this ElementNode node)
-        {
-            if (node == null)
-            {
-                return;
-            }
-
-            var children = node.Children().Cast<ElementNode>().ToList();
-            foreach (var child in children)
-            {
-                RemoveNullChildren(child);
-            }
-
-            if (!node.Children().Any() && node.Value == null && !Enum.TryParse<ResourceType>(node.InstanceType, true, out _))
-            {
-                node.Parent.Remove(node);
-                return;
-            }
-        }
-
-        public static void AddSecurityTag(this ElementNode node, ProcessResult result, IStructureDefinitionSummaryProvider provider)
-        {
-            if (node == null)
-            {
-                return;
-            }
-
-            ElementNode metaNode = node.GetMeta();
-            Meta meta = metaNode?.ToPoco<Meta>() ?? new Meta();
-
-            if (result.IsRedacted && !meta.Security.Any(x =>
-                string.Equals(x.Code, SecurityLabels.REDACT.Code, StringComparison.InvariantCultureIgnoreCase)))
-            {
-                meta.Security.Add(SecurityLabels.REDACT);
-            }
-
-            if (result.IsAbstracted && !meta.Security.Any(x =>
-                string.Equals(x.Code, SecurityLabels.ABSTRED.Code, StringComparison.InvariantCultureIgnoreCase)))
-            {
-                meta.Security.Add(SecurityLabels.ABSTRED);
-            }
-
-            if (result.IsPerturbed && !meta.Security.Any(x =>
-                string.Equals(x.Code, SecurityLabels.PERTURBED.Code, StringComparison.InvariantCultureIgnoreCase)))
-            {
-                meta.Security.Add(SecurityLabels.PERTURBED);
-            }
-
-            ElementNode newMetaNode = ElementNode.FromElement(meta.ToTypedElement());
-            if (metaNode == null)
-            {
-                node.Add(provider, newMetaNode);
-            }
-            else 
-            {
-                node.Replace(provider, metaNode, newMetaNode);
-            }
-        }
     }
 }
