+++ src/Fhir.Anonymizer.Core/Visitors/AbstractElementNodeVisitor.cs
@@ -7,26 +7,6 @@ namespace Fhir.Anonymizer.Core.Visitors
 {
     public abstract class AbstractElementNodeVisitor
     {
-        public virtual bool PreVisitContainedNode(ElementNode node)
-        {
-            return true;
-        }
-
-        public virtual void PostVisitContainedNode(ElementNode node)
-        {
-            
-        }
-
-        public virtual bool PreVisitBundleEntryNode(ElementNode node)
-        {
-            return true;
-        }
-
-        public virtual void PostVisitBundleEntryNode(ElementNode node)
-        {
-
-        }
-
         public virtual bool Visit(ElementNode node)
         {
             return true;
